# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management API",
        "description": "This API handles the management of customer orders",
        "contact": {
            "email": "aloshmohammad2001@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a list of customers",
                "description": "Returns a paginated list of customers with their details",
                "operationId": "8acc5e9f5561eec1edbaf6a68f600e82",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "Id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "FullName": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "Country": {
                                                                "type": "string",
                                                                "example": "USA"
                                                            },
                                                            "PhoneNumber": {
                                                                "type": "string",
                                                                "example": "123-456-7890"
                                                            },
                                                            "Email": {
                                                                "type": "string",
                                                                "example": "john.doe@example.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "description": "Creates a new customer with the provided details",
                "operationId": "428d5533c8386f91d2baf4de69f01468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "FullName",
                                    "Email",
                                    "Country",
                                    "PhoneNumber",
                                    "Address"
                                ],
                                "properties": {
                                    "FullName": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "Email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "Country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "PhoneNumber": {
                                        "type": "string",
                                        "example": "123-456-7890"
                                    },
                                    "Address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "Note": {
                                        "type": "string",
                                        "example": "Preferred customer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "Id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "FullName": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "Email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "Country": {
                                                    "type": "string",
                                                    "example": "USA"
                                                },
                                                "PhoneNumber": {
                                                    "type": "string",
                                                    "example": "123-456-7890"
                                                },
                                                "Address": {
                                                    "type": "string",
                                                    "example": "123 Main St"
                                                },
                                                "Note": {
                                                    "type": "string",
                                                    "example": "Preferred customer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/customers/search": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Search for customers",
                "description": "Search for customers by various fields such as name, email, country, etc.",
                "operationId": "36cbc04592be7f0051a44abe99503238",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query for customers",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "search": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "Id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "SeqNumber": {
                                                                "type": "string",
                                                                "example": "S123456"
                                                            },
                                                            "FullName": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "Country": {
                                                                "type": "string",
                                                                "example": "USA"
                                                            },
                                                            "PhoneNumber": {
                                                                "type": "string",
                                                                "example": "123-456-7890"
                                                            },
                                                            "Email": {
                                                                "type": "string",
                                                                "example": "john.doe@example.com"
                                                            },
                                                            "Address": {
                                                                "type": "string",
                                                                "example": "123 Main St"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}